pke: split create_cluster.go
