+++ internal/providers/pke/pkeworkflow/cluster.go
@@ -1,0 +1,29 @@
+package pkeworkflow
+
+type Clusters interface {
+	GetCluster(ctx context.Context, id uint) (Cluster, error)
+}
+
+type Cluster interface {
+	GetID() uint
+	GetUID() string
+	GetName() string
+	GetOrganizationId() uint
+	UpdateStatus(string, string) error
+	GetNodePools() []NodePool
+}
+
+type AWSCluster interface {
+	GetAWSClient() (*session.Session, error)
+	GetBootstrapCommand(string, string, string) (string, error)
+	SaveNetworkCloudProvider(string, string, []string) error
+}
+
+type NodePool struct {
+	Name     string
+	MinCount int
+	MaxCount int
+	Count    int
+	Master   bool
+	Worker   bool
+}
