+++ internal/providers/pke/pkeworkflow/create_cluster.go
@@ -24,34 +24,6 @@ import (
 	"go.uber.org/zap"
 )
 
-type Clusters interface {
-	GetCluster(ctx context.Context, id uint) (Cluster, error)
-}
-
-type Cluster interface {
-	GetID() uint
-	GetUID() string
-	GetName() string
-	GetOrganizationId() uint
-	UpdateStatus(string, string) error
-	GetNodePools() []NodePool
-}
-
-type AWSCluster interface {
-	GetAWSClient() (*session.Session, error)
-	GetBootstrapCommand(string, string, string) (string, error)
-	SaveNetworkCloudProvider(string, string, []string) error
-}
-
-type NodePool struct {
-	Name     string
-	MinCount int
-	MaxCount int
-	Count    int
-	Master   bool
-	Worker   bool
-}
-
 const CreateClusterWorkflowName = "pke-create-cluster"
 
 type CreateClusterWorkflowInput struct {
